Convert from spaces to tabs for TodoMVC compliance
