+++ labs/architecture-examples/montage/assets/app.css
@@ -1,3 +1,3 @@
 #montage-todomvc .visible {
-    display: block;
+	display: block;
 }
