+++ labs/architecture-examples/montage/bower.json
