+++ labs/architecture-examples/montage/core/todo.js
@@ -2,25 +2,25 @@ var Montage = require('montage').Montage;
 
 exports.Todo = Montage.specialize({
 
-    constructor: {
-        value: function Todo() {
-            this.super();
-        }
-    },
+	constructor: {
+		value: function Todo() {
+			this.super();
+		}
+	},
 
-    initWithTitle: {
-        value: function (title) {
-            this.title = title;
-            return this;
-        }
-    },
+	initWithTitle: {
+		value: function (title) {
+			this.title = title;
+			return this;
+		}
+	},
 
-    title: {
-        value: null
-    },
+	title: {
+		value: null
+	},
 
-    completed: {
-        value: false
-    }
+	completed: {
+		value: false
+	}
 
 });
