+++ labs/architecture-examples/montage/index.html
@@ -1,27 +1,27 @@
 <!doctype html>
 <html lang="en" data-framework="montage" id="montage-todomvc">
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <title>Montage • TodoMVC</title>
+	<head>
+		<meta charset="utf-8">
+		<meta http-equiv="X-UA-Compatible" content="IE=edge">
+		<title>Montage • TodoMVC</title>
 
-        <link rel="stylesheet" href="bower_components/todomvc-common/base.css">
-        <link rel="stylesheet" href="assets/app.css">
-    </head>
-    <body>
-        <div data-montage-id="todo-container" id="todo-container"></div>
+		<link rel="stylesheet" href="bower_components/todomvc-common/base.css">
+		<link rel="stylesheet" href="assets/app.css">
+	</head>
+	<body>
+		<div data-montage-id="todo-container" id="todo-container"></div>
 
-        <script src="bower_components/todomvc-common/base.js"></script>
-        <script src="index.html.bundle-0.js" data-montage-location="packages/montage@6364dae/" data-promise-location="packages/q@2847ee2/" data-montage-hash="6364dae" data-promise-hash="2847ee2" data-application-hash="37bb2cd"></script>
-        <script type="text/montage-serialization">
-            {
-                "owner": {
-                    "prototype": "montage/ui/loader.reel",
-                    "properties": {
-                        "element": {"#": "todo-container"}
-                    }
-                }
-            }
-        </script>
-    </body>
+		<script src="bower_components/todomvc-common/base.js"></script>
+		<script src="index.html.bundle-0.js" data-montage-location="packages/montage@6364dae/" data-promise-location="packages/q@2847ee2/" data-montage-hash="6364dae" data-promise-hash="2847ee2" data-application-hash="6607c26"></script>
+		<script type="text/montage-serialization">
+			{
+				"owner": {
+					"prototype": "montage/ui/loader.reel",
+					"properties": {
+						"element": {"#": "todo-container"}
+					}
+				}
+			}
+		</script>
+	</body>
 </html>
