+++ labs/architecture-examples/montage/ui/main.reel/main.css
@@ -1,5 +1,5 @@
 #main,
 #footer,
 #clear-completed-container {
-    display: none;
+	display: none;
 }
