+++ labs/architecture-examples/montage/ui/main.reel/main.js
@@ -7,172 +7,172 @@ var LOCAL_STORAGE_KEY = 'todos-montage';
 
 exports.Main = Component.specialize({
 
-    _newTodoForm: {
-        value: null
-    },
+	_newTodoForm: {
+		value: null
+	},
 
-    _newTodoInput: {
-        value: null
-    },
+	_newTodoInput: {
+		value: null
+	},
 
-    todoListController: {
-        value: null
-    },
+	todoListController: {
+		value: null
+	},
 
-    constructor: {
-        value: function Main() {
-            this.todoListController = new RangeController();
-            this.addPathChangeListener('todos.every{completed}', this, 'handleTodosCompletedChanged');
+	constructor: {
+		value: function Main() {
+			this.todoListController = new RangeController();
+			this.addPathChangeListener('todos.every{completed}', this, 'handleTodosCompletedChanged');
 
-            this.defineBindings({
-                'todos': {'<-': 'todoListController.organizedContent'},
-                'todosLeft': {'<-': 'todos.filter{!completed}'},
-                'todosCompleted': {'<-': 'todos.filter{completed}'}
-            });
-        }
-    },
+			this.defineBindings({
+				'todos': {'<-': 'todoListController.organizedContent'},
+				'todosLeft': {'<-': 'todos.filter{!completed}'},
+				'todosCompleted': {'<-': 'todos.filter{completed}'}
+			});
+		}
+	},
 
-    templateDidLoad: {
-        value: function () {
-            this.load();
-        }
-    },
+	templateDidLoad: {
+		value: function () {
+			this.load();
+		}
+	},
 
-    load: {
-        value: function () {
-            if (localStorage) {
-                var todoSerialization = localStorage.getItem(LOCAL_STORAGE_KEY);
+	load: {
+		value: function () {
+			if (localStorage) {
+				var todoSerialization = localStorage.getItem(LOCAL_STORAGE_KEY);
 
-                if (todoSerialization) {
-                    var deserializer = new Deserializer(),
-                        self = this;
+				if (todoSerialization) {
+					var deserializer = new Deserializer(),
+						self = this;
 
-                    deserializer.init(todoSerialization, require)
-                    .deserializeObject()
-                    .then(function (todos) {
-                        self.todoListController.content = todos;
-                    }).fail(function (error) {
-                        console.error('Could not load saved tasks.');
-                        console.debug('Could not deserialize', todoSerialization);
-                        console.log(error.stack);
-                    });
-                }
-            }
-        }
-    },
+					deserializer.init(todoSerialization, require)
+					.deserializeObject()
+					.then(function (todos) {
+						self.todoListController.content = todos;
+					}).fail(function (error) {
+						console.error('Could not load saved tasks.');
+						console.debug('Could not deserialize', todoSerialization);
+						console.log(error.stack);
+					});
+				}
+			}
+		}
+	},
 
-    save: {
-        value: function () {
-            if (localStorage) {
-                var todos = this.todoListController.content,
-                    serializer = new Serializer().initWithRequire(require);
+	save: {
+		value: function () {
+			if (localStorage) {
+				var todos = this.todoListController.content,
+					serializer = new Serializer().initWithRequire(require);
 
-                localStorage.setItem(LOCAL_STORAGE_KEY, serializer.serializeObject(todos));
-            }
-        }
-    },
+				localStorage.setItem(LOCAL_STORAGE_KEY, serializer.serializeObject(todos));
+			}
+		}
+	},
 
-    enterDocument: {
-        value: function (firstTime) {
-            if (firstTime) {
-                this._newTodoForm.identifier = 'newTodoForm';
-                this._newTodoForm.addEventListener('submit', this, false);
+	enterDocument: {
+		value: function (firstTime) {
+			if (firstTime) {
+				this._newTodoForm.identifier = 'newTodoForm';
+				this._newTodoForm.addEventListener('submit', this, false);
 
-                this.addEventListener('destroyTodo', this, true);
+				this.addEventListener('destroyTodo', this, true);
 
-                window.addEventListener('beforeunload', this, true);
-            }
-        }
-    },
+				window.addEventListener('beforeunload', this, true);
+			}
+		}
+	},
 
-    captureDestroyTodo: {
-        value: function (evt) {
-            this.destroyTodo(evt.detail.todo);
-        }
-    },
+	captureDestroyTodo: {
+		value: function (evt) {
+			this.destroyTodo(evt.detail.todo);
+		}
+	},
 
-    createTodo: {
-        value: function (title) {
-            var todo = new Todo().initWithTitle(title);
-            this.todoListController.add(todo);
-            return todo;
-        }
-    },
+	createTodo: {
+		value: function (title) {
+			var todo = new Todo().initWithTitle(title);
+			this.todoListController.add(todo);
+			return todo;
+		}
+	},
 
-    destroyTodo: {
-        value: function (todo) {
-            this.todoListController.delete(todo);
-            return todo;
-        }
-    },
+	destroyTodo: {
+		value: function (todo) {
+			this.todoListController.delete(todo);
+			return todo;
+		}
+	},
 
-    _allCompleted: {
-        value: null
-    },
+	_allCompleted: {
+		value: null
+	},
 
-    allCompleted: {
-        get: function () {
-            return this._allCompleted;
-        },
-        set: function (value) {
-            this._allCompleted = value;
-            this.todoListController.organizedContent.forEach(function (member) {
-                member.completed = value;
-            });
-        }
-    },
+	allCompleted: {
+		get: function () {
+			return this._allCompleted;
+		},
+		set: function (value) {
+			this._allCompleted = value;
+			this.todoListController.organizedContent.forEach(function (member) {
+				member.completed = value;
+			});
+		}
+	},
 
-    todos: {
-        value: null
-    },
+	todos: {
+		value: null
+	},
 
-    todosLeft: {
-        value: null
-    },
+	todosLeft: {
+		value: null
+	},
 
-    todosCompleted: {
-        value: null
-    },
+	todosCompleted: {
+		value: null
+	},
 
-    // Handlers
+	// Handlers
 
-    handleNewTodoFormSubmit: {
-        value: function (evt) {
-            evt.preventDefault();
+	handleNewTodoFormSubmit: {
+		value: function (evt) {
+			evt.preventDefault();
 
-            var title = this._newTodoInput.value.trim();
+			var title = this._newTodoInput.value.trim();
 
-            if (title === '') {
-                return;
-            }
+			if (title === '') {
+				return;
+			}
 
-            this.createTodo(title);
-            this._newTodoInput.value = null;
-        }
-    },
+			this.createTodo(title);
+			this._newTodoInput.value = null;
+		}
+	},
 
-    handleTodosCompletedChanged: {
-        value: function (value) {
-            this._allCompleted = value;
-            this.dispatchOwnPropertyChange('allCompleted', value);
-        }
-    },
+	handleTodosCompletedChanged: {
+		value: function (value) {
+			this._allCompleted = value;
+			this.dispatchOwnPropertyChange('allCompleted', value);
+		}
+	},
 
-    handleClearCompletedButtonAction: {
-        value: function () {
-            var completedTodos = this.todoListController.organizedContent.filter(function (todo) {
-                return todo.completed;
-            });
+	handleClearCompletedButtonAction: {
+		value: function () {
+			var completedTodos = this.todoListController.organizedContent.filter(function (todo) {
+				return todo.completed;
+			});
 
-            if (completedTodos.length > 0) {
-                this.todoListController.deleteEach(completedTodos);
-            }
-        }
-    },
+			if (completedTodos.length > 0) {
+				this.todoListController.deleteEach(completedTodos);
+			}
+		}
+	},
 
-    captureBeforeunload: {
-        value: function () {
-            this.save();
-        }
-    }
+	captureBeforeunload: {
+		value: function () {
+			this.save();
+		}
+	}
 });
